Use correct function to get normalized Vector
The Vector:Normalize() function now modifies the Vector, instead of
simply returning the normalized Vector. This means that
Vector:Normalize() no longer returns a value which is a problem since
it is being multiplied. The new function Vector:GetNormalized() works
like the old Vector:Normalize() does, returning the normalized vector,
so it works as a replacement.